Create useEntries hook
I'm not pretty sure if this is the best way to implement this hook, but it's working for now.
